refactor: Move valid file extensions to workspace.constants
